refactor. store 파일수정 1. StoreService 인터페이스 구현하여 StoreServiceImpl에서 실제 구현완료 2. APiResponse result.getData()없앰 컨트롤러에서 result 하나만 넣어주면됨
